+++ src/jloptions.c
@@ -4,6 +4,7 @@
 #include <errno.h>
 
 #include "julia.h"
+#include "julia_internal.h"
 
 #include <unistd.h>
 #include <getopt.h>
@@ -73,7 +74,7 @@ JL_DLLEXPORT void jl_init_options(void)
                         NULL, // output-o
                         NULL, // output-asm
                         NULL, // output-ji
-                        NULL,    // output-code_coverage
+                        NULL, // output-code_coverage
                         0, // incremental
                         0, // image_file_specified
                         JL_OPTIONS_WARN_SCOPE_ON,  // ambiguous scope warning
@@ -85,97 +86,106 @@ JL_DLLEXPORT void jl_init_options(void)
     jl_options_initialized = 1;
 }
 
-static const char usage[] = "julia [switches] -- [programfile] [args...]\n";
+static const char usage[] = "\n    julia [switches] -- [programfile] [args...]\n\n";
 static const char opts[]  =
-    " -v, --version             Display version information\n"
-    " -h, --help                Print this message (--help-hidden for more)\n"
-    " --help-hidden             Uncommon options not shown by `-h`\n\n"
+    "Switches (a '*' marks the default value, if applicable):\n\n"
+    " -v, --version              Display version information\n"
+    " -h, --help                 Print this message (--help-hidden for more)\n"
+    " --help-hidden              Uncommon options not shown by `-h`\n\n"
 
     // startup options
-    " --project[={<dir>|@.}]    Set <dir> as the home project/environment\n"
-    " -J, --sysimage <file>     Start up with the given system image file\n"
-    " -H, --home <dir>          Set location of `julia` executable\n"
-    " --startup-file={yes|no}   Load `~/.julia/config/startup.jl`\n"
-    " --handle-signals={yes|no} Enable or disable Julia's default signal handlers\n"
-    " --sysimage-native-code={yes|no}\n"
-    "                           Use native code from system image if available\n"
-    " --compiled-modules={yes|no}\n"
-    "                           Enable or disable incremental precompilation of modules\n\n"
+    " --project[={<dir>|@.}]     Set <dir> as the home project/environment\n"
+    " -J, --sysimage <file>      Start up with the given system image file\n"
+    " -H, --home <dir>           Set location of `julia` executable\n"
+    " --startup-file={yes*|no}   Load `JULIA_DEPOT_PATH/config/startup.jl`; if `JULIA_DEPOT_PATH`\n"
+    "                            environment variable is unset, load `~/.julia/config/startup.jl`\n"
+    " --handle-signals={yes*|no} Enable or disable Julia's default signal handlers\n"
+    " --sysimage-native-code={yes*|no}\n"
+    "                            Use native code from system image if available\n"
+    " --compiled-modules={yes*|no}\n"
+    "                            Enable or disable incremental precompilation of modules\n\n"
 
     // actions
-    " -e, --eval <expr>         Evaluate <expr>\n"
-    " -E, --print <expr>        Evaluate <expr> and display the result\n"
-    " -L, --load <file>         Load <file> immediately on all processors\n\n"
+    " -e, --eval <expr>          Evaluate <expr>\n"
+    " -E, --print <expr>         Evaluate <expr> and display the result\n"
+    " -L, --load <file>          Load <file> immediately on all processors\n\n"
 
     // parallel options
-    " -t, --threads {N|auto}    Enable N threads; \"auto\" currently sets N to the number of local\n"
-    "                           CPU threads but this might change in the future\n"
-    " -p, --procs {N|auto}      Integer value N launches N additional local worker processes\n"
-    "                           \"auto\" launches as many workers as the number of local CPU threads (logical cores)\n"
-    " --machine-file <file>     Run processes on hosts listed in <file>\n\n"
+    " -t, --threads {N|auto}     Enable N threads; \"auto\" tries to infer a useful default number\n"
+    "                            of threads to use but the exact behavior might change in the future.\n"
+    "                            Currently, \"auto\" uses the number of CPUs assigned to this julia\n"
+    "                            process based on the OS-specific affinity assignment interface, if\n"
+    "                            supported (Linux and Windows). If this is not supported (macOS) or\n"
+    "                            process affinity is not configured, it uses the number of CPU\n"
+    "                            threads.\n"
+    " -p, --procs {N|auto}       Integer value N launches N additional local worker processes\n"
+    "                            \"auto\" launches as many workers as the number of local CPU threads (logical cores)\n"
+    " --machine-file <file>      Run processes on hosts listed in <file>\n\n"
 
     // interactive options
-    " -i                        Interactive mode; REPL runs and isinteractive() is true\n"
-    " -q, --quiet               Quiet startup: no banner, suppress REPL warnings\n"
-    " --banner={yes|no|auto}    Enable or disable startup banner\n"
-    " --color={yes|no|auto}     Enable or disable color text\n"
-    " --history-file={yes|no}   Load or save history\n\n"
+    " -i                         Interactive mode; REPL runs and `isinteractive()` is true\n"
+    " -q, --quiet                Quiet startup: no banner, suppress REPL warnings\n"
+    " --banner={yes|no|auto*}    Enable or disable startup banner\n"
+    " --color={yes|no|auto*}     Enable or disable color text\n"
+    " --history-file={yes*|no}   Load or save history\n\n"
 
     // error and warning options
-    " --depwarn={yes|no|error}  Enable or disable syntax and method deprecation warnings (\"error\" turns warnings into errors)\n"
-    " --warn-overwrite={yes|no} Enable or disable method overwrite warnings\n"
-    " --warn-scope={yes|no}     Enable or disable warning for ambiguous top-level scope\n\n"
+    " --depwarn={yes|no*|error}  Enable or disable syntax and method deprecation warnings (`error` turns warnings into errors)\n"
+    " --warn-overwrite={yes|no*} Enable or disable method overwrite warnings\n"
+    " --warn-scope={yes*|no}     Enable or disable warning for ambiguous top-level scope\n\n"
 
     // code generation options
-    " -C, --cpu-target <target> Limit usage of CPU features up to <target>; set to \"help\" to see the available options\n"
-    " -O, --optimize={0,1,2,3}  Set the optimization level (default level is 2 if unspecified or 3 if used without a level)\n"
-    " --min-optlevel={0,1,2,3}  Set a lower bound on the optimization level (default is 0)\n"
-    " -g, -g <level>            Enable or set the level of debug info generation"
+    " -C, --cpu-target <target>  Limit usage of CPU features up to <target>; set to `help` to see the available options\n"
+    " -O, --optimize={0,1,2*,3}  Set the optimization level (level 3 if `-O` is used without a level)\n"
+    " --min-optlevel={0*,1,2,3}  Set a lower bound on the optimization level\n"
 #ifdef JL_DEBUG_BUILD
-        " (default level for julia-debug is 2 if unspecified or if used without a level)\n"
+        " -g [{0,1,2*}]              Set the level of debug info generation in the julia-debug build\n"
 #else
-        " (default level is 1 if unspecified or 2 if used without a level)\n"
+        " -g [{0,1*,2}]              Set the level of debug info generation (level 2 if `-g` is used without a level)\n"
 #endif
-    " --inline={yes|no}         Control whether inlining is permitted, including overriding @inline declarations\n"
-    " --check-bounds={yes|no|auto}\n"
-    "                           Emit bounds checks always, never, or respect @inbounds declarations\n"
+    " --inline={yes*|no}         Control whether inlining is permitted, including overriding @inline declarations\n"
+    " --check-bounds={yes|no|auto*}\n"
+    "                            Emit bounds checks always, never, or respect @inbounds declarations\n"
 #ifdef USE_POLLY
-    " --polly={yes|no}          Enable or disable the polyhedral optimizer Polly (overrides @polly declaration)\n"
+    " --polly={yes*|no}          Enable or disable the polyhedral optimizer Polly (overrides @polly declaration)\n"
 #endif
-    " --math-mode={ieee,fast}   Disallow or enable unsafe floating point optimizations (overrides @fastmath declaration)\n\n"
+    " --math-mode={ieee,fast}    Disallow or enable unsafe floating point optimizations (overrides @fastmath declaration)\n\n"
 
     // instrumentation options
-    " --code-coverage={none|user|all}, --code-coverage\n"
-    "                           Count executions of source lines (omitting setting is equivalent to \"user\")\n"
+    " --code-coverage[={none*|user|all}]\n"
+    "                            Count executions of source lines (omitting setting is equivalent to `user`)\n"
     " --code-coverage=tracefile.info\n"
-    "                           Append coverage information to the LCOV tracefile (filename supports format tokens).\n"
+    "                            Append coverage information to the LCOV tracefile (filename supports format tokens)\n"
 // TODO: These TOKENS are defined in `runtime_ccall.cpp`. A more verbose `--help` should include that list here.
-    " --track-allocation={none|user|all}, --track-allocation\n"
-    "                           Count bytes allocated by each source line (omitting setting is equivalent to \"user\")\n"
-    " --bug-report=KIND         Launch a bug report session. It can be used to start a REPL, run a script, or evaluate\n"
-    "                           expressions. It first tries to use BugReporting.jl installed in current environment and\n"
-    "                           fallbacks to the latest compatible BugReporting.jl if not. For more information, see\n"
-    "                           --bug-report=help.\n\n"
+    " --track-allocation[={none*|user|all}]\n"
+    "                            Count bytes allocated by each source line (omitting setting is equivalent to `user`)\n"
+    " --bug-report=KIND          Launch a bug report session. It can be used to start a REPL, run a script, or evaluate\n"
+    "                            expressions. It first tries to use BugReporting.jl installed in current environment and\n"
+    "                            fallbacks to the latest compatible BugReporting.jl if not. For more information, see\n"
+    "                            --bug-report=help.\n\n"
 ;
 
 static const char opts_hidden[]  =
+    "Switches (a '*' marks the default value, if applicable):\n\n"
     // code generation options
-    " --compile={yes|no|all|min}Enable or disable JIT compiler, or request exhaustive or minimal compilation\n"
+    " --compile={yes*|no|all|min}\n"
+    "                          Enable or disable JIT compiler, or request exhaustive or minimal compilation\n\n"
 
     // compiler output options
-    " --output-o name           Generate an object file (including system image data)\n"
-    " --output-ji name          Generate a system image data file (.ji)\n"
-    " --strip-metadata          Remove docstrings and source location info from system image\n"
-    " --strip-ir                Remove IR (intermediate representation) of compiled functions\n"
+    " --output-o <name>        Generate an object file (including system image data)\n"
+    " --output-ji <name>       Generate a system image data file (.ji)\n"
+    " --strip-metadata         Remove docstrings and source location info from system image\n"
+    " --strip-ir               Remove IR (intermediate representation) of compiled functions\n\n"
 
     // compiler debugging (see the devdocs for tips on using these options)
-    " --output-unopt-bc name    Generate unoptimized LLVM bitcode (.bc)\n"
-    " --output-bc name          Generate LLVM bitcode (.bc)\n"
-    " --output-asm name         Generate an assembly file (.s)\n"
-    " --output-incremental=no   Generate an incremental output file (rather than complete)\n"
+    " --output-unopt-bc <name> Generate unoptimized LLVM bitcode (.bc)\n"
+    " --output-bc <name>       Generate LLVM bitcode (.bc)\n"
+    " --output-asm <name>      Generate an assembly file (.s)\n"
+    " --output-incremental={yes|no*}\n"
+    "                          Generate an incremental output file (rather than complete)\n"
     " --trace-compile={stderr,name}\n"
-    "                           Print precompile statements for methods compiled during execution or save to a path\n\n"
-    " --image-codegen           Force generate code in imaging mode\n"
+    "                          Print precompile statements for methods compiled during execution or save to a path\n"
+    " --image-codegen          Force generate code in imaging mode\n"
 ;
 
 JL_DLLEXPORT void jl_parse_opts(int *argcp, char ***argvp)
@@ -236,7 +246,7 @@ JL_DLLEXPORT void jl_parse_opts(int *argcp, char ***argvp)
         { "bug-report",      required_argument, 0, opt_bug_report },
         { "sysimage",        required_argument, 0, 'J' },
         { "sysimage-native-code", required_argument, 0, opt_sysimage_native_code },
-        { "compiled-modules",    required_argument, 0, opt_compiled_modules },
+        { "compiled-modules",required_argument, 0, opt_compiled_modules },
         { "cpu-target",      required_argument, 0, 'C' },
         { "procs",           required_argument, 0, 'p' },
         { "threads",         required_argument, 0, 't' },
@@ -441,7 +451,7 @@ restart_switch:
         case 'p': // procs
             errno = 0;
             if (!strcmp(optarg,"auto")) {
-                jl_options.nprocs = jl_cpu_threads();
+                jl_options.nprocs = jl_effective_threads();
             }
             else {
                 long nprocs = strtol(optarg, &endptr, 10);